Display PSTN/SIP provided incoming name
The current Dialer implementation strips the provided display name and
only displays the number unless the user has a contact in which case
it displays the contact name.  For SIP, in particular, where the
display name might be useful, this is suboptimal.  Update the
CallerInfoAsyncQuery to pass the display name if a contact name
doesn't exist.

Also, since the display name may not be definitive, update the status
bar notifier to include the number in brackets after the name if the
name isn't a contact name.

Change-Id: I204faf76a6827ff1df5cd9a3d3f496556d7c4027
Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>